Moved redux containers folders + testing
Redux containers are now in their own folder
Added testing for reducers and one of the containers
